chore: replace Karma+Jasmine by Jest and add some examples unit tests
Karma is the historical solution for running unit tests in an AngularJS app.
However, it is now quite outdated and unpleasant to use. Jest is another JS testing framework with both better performances, dev experience.

The command to run the unit tests is still the same, npm test, but Jest adds some useful options like:

- Running test from a single test file: npm test -- --testPathPattern=analytics.service.spec.ts
- A top-of-the-art watch mode: npm test -- --testPathPattern=analytics.service.spec.ts --watch
- A perfectly working integration with VSCode and Jetbrains IDE

I also added simple test examples for a service and a component.